Renamed mapping to quimb and added new unit test for tensors with phases
